Expose function to capture EXCEPTION_POINTERS
We need a way to manually report crashes to Sentry or exceptions that
will crash the application. Sentry does this by using a struct which
contains an EXCEPTION_POINTERS to try to send it into the web service.
We can expose this function and add a new button in our test application
which creates an empty EXCEPTION_POINTERS to send into the crash system.